refactor(vrc25): seperate vrc25 package to keeps clean
